% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bca_ppi}
\alias{bca_ppi}
\title{A protein-protein interactions (PPIs) dataset of breast cancer from CancerNet database. This PPIs dataset contains 143751 edges (interactions) among 10597 nodes (proteins).}
\description{
A protein-protein interactions dataset of breast cancer
}
\details{
An igraph graph object containing 143751 edges (interactions) among 10597 nodes (proteins).
}
\references{
Meng X, Wang J, Yuan C. et al. CancerNet: a database for decoding multilevel molecular interactions across diverse cancer types. Oncogenesis 4, e177 (2015).
}
\keyword{input-datasets}
