% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netva.R
\name{compactness}
\alias{compactness}
\title{Calculates compactness of network based on the presence of all maximal 3-node cliques in complete network
Or
Determine network compactness after removal of a node of interest}
\usage{
compactness(v = NULL, net, p = 0.1)
}
\arguments{
\item{v}{A node/protein name which will be removed (by default NULL, i.e., none of nodes will be removed and the compactness value
will be of complete network).}

\item{net}{An igraph graph object.}

\item{p}{A penalty term which measures the inaccuracy of the network interaction. The default value of p is set to 0.1.}
}
\value{
A single numeric value.
}
\description{
Calculates compactness of network based on the presence of all maximal 3-node cliques in complete network
Or
Determine network compactness after removal of a node of interest
}
