% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectHB.R, R/detectHub-Bot.R
\name{detectHubs}
\alias{detectHubs}
\title{Identify all possible hubs based on the pareto principle of Eighty-twenty rule (by default) for a given network}
\usage{
detectHubs(net, p = 20)

detectHubs(net, p = 20)
}
\arguments{
\item{net}{An igraph graph object.}

\item{p}{Value of percent of nodes/proteins to be considered for the determination of degree cutoff. By default value is 20\%. It is an optional parameter.}

\item{method}{Method to identify hub nodes. Currently this function supports only the method of Eighty-twenty principle (ETP).}

\item{validate}{Logical, TRUE or FALSE, whether to validate identified hubs by rewiring of a given percentage of edges.}

\item{perturb}{Percentage value to rewire edges.}

\item{iter}{Number of iterations to perform rewiring and construction of new rewired networks. The default value is round(100/perturb).}

\item{ng}{Number of new graphs/networks to be constructed. The default value is the value of iter.}
}
\value{
A numeric vector containing all possible hubs with hub proteins' names and corresponding degree values as identified in the given network.

A numeric vector containing all possible hubs with hub proteins' names and corresponding degree values as identified in the given network.
}
\description{
Identify all possible hubs based on the pareto principle of Eighty-twenty rule (by default) for a given network

Identify all possible hubs based on pareto principle of 80:20 rule (by default) for a given network
}
