% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectINs.R
\name{detectINs}
\alias{detectINs}
\title{Identify all possible Influential nodes (INs) based on pareto principle of 80:20 rule (by default) for a given network}
\usage{
detectINs(net, p = 20)
}
\arguments{
\item{net}{An igraph graph object.}

\item{p}{Value of percent of nodes to be considered for the determination of EVC/EVC+ cutoff. By default value is 20\%.}
}
\value{
A list containing two vectors: (1) ins.evc - all possible Influential nodes based on EVC, (2) ins.evcplus - all possible
Influential nodes based on EVC+ as identified in the given network.
}
\description{
Identify all possible Influential nodes (INs) based on pareto principle of 80:20 rule (by default) for a given network
}
