% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evc.R
\name{evc}
\alias{evc}
\title{Escape velocity centrality (normal and extended version)}
\usage{
evc(g, alpha = 1, mode = "all")
}
\arguments{
\item{g}{An igraph graph object.}

\item{alpha}{A tunable factor with value between 0.1 to 1. The default value of alpha is 1.}

\item{mode}{The type of the core in a graph. Character constant with possible values - ‘in’: in-cores are computed, ‘out’:
out-cores are computed, ‘all’: the corresponding undirected graph is considered. This argument is only for directed
graphs.}
}
\value{
A list of two vectors: (1) evc, containing the values of EVC and
(2) evc.plus, the values of EVC+ of each vertex in the network.
}
\description{
Calculates escape velocity centrality (EVC, evc) and extended escape velocity centrality (EVC+, evc.plus)
of all vertices in the network under study \link{Ullah et al., 2022}.
}
\references{
Ullah A. et al. (2022). Escape velocity centrality: escape influence-based key nodes identification in complex networks, Applied Intelligence, 1-19.
}
