% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bcppi.R
\name{load_bcppi}
\alias{load_bcppi}
\title{Loads protein-protein interactions data of breast cancer}
\usage{
load_bcppi()
}
\value{
A dataframe containing protein-protein interactions data of breast cancer.
}
\description{
Source: CancerNet database (http://bis.zju.edu.cn/CancerNet)
Reference: Meng, X., Wang, J., Yuan, C. et al. CancerNet: a database for decoding multilevel molecular interactions across diverse cancer types. Oncogenesis 4, e177 (2015). https://doi.org/10.1038/oncsis.2015.40
}
