% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netva.R
\name{netva.mc}
\alias{netva.mc}
\title{Perform network vulnerability analysis on multiple cores}
\usage{
netva.mc(vl, net, type, ncore)
}
\arguments{
\item{vl}{A vector of node or protein names}

\item{net}{Takes input an igraph graph object}

\item{type}{A keyword either "full" or "cluster" for consideration of complete network or only the largest cluster}

\item{ncore}{Number of cores users want to use for the computation}
}
\value{
A dataframe having number of rows equal to the number of proteins in consideration and ten columns
}
\description{
This function performs deletion of nodes one by one and calculation of average node connectivity, average path length, articulation
point, network centralization, network density, network diameter, transitivity (clustering coefficient), betweenness and heterogeneity
of resultant networks using multiple cores. It is a wrapper around netva() utilizing parallel computing features of R. It is only Linux/
macOS based machine compatible.
}
