% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netva.R
\name{netva}
\alias{netva}
\title{Perform network vulnerability analysis for a given network}
\usage{
netva(v, net, type)
}
\arguments{
\item{v}{Name of any one node/protein to be deleted, or a character vector of names to delete more than one nodes or proteins.}

\item{net}{An igraph graph object.}

\item{type}{A keyword either "full" or "cluster" for consideration of complete network or only the largest cluster.}
}
\value{
A numeric vector of nine elements corresponding to nine different topological properties.
}
\description{
This function performs deletion of one or more than one nodes and calculates nine different topological properties such as average node
connectivity, average path length, articulation point, network centralization, network density, network diameter, transitivity
(clustering coefficient), betweenness and heterogeneity for the resultant network.
}
