% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netva.R
\name{netva.sc}
\alias{netva.sc}
\title{Perform network vulnerability analysis using single core only}
\usage{
netva.sc(vl, net, type)
}
\arguments{
\item{vl}{A vector of node or protein names}

\item{net}{Takes input an igraph graph object}

\item{type}{A keyword either "full" or "cluster" for consideration of complete network or only the largest cluster}
}
\value{
A dataframe having number of rows equal to the number of proteins in consideration and ten columns
}
\description{
This function performs deletion of nodes one by one and calculation of average node connectivity, average path length, articulation
point, network centralization, network density, network diameter, transitivity (clustering coefficient), betweenness and heterogeneity
of resultant networks. It is a wrapper around netva() for performing network vulnerability analysis on any machines such as Linux,
macOS or Windows utilizing single core.
}
