% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{netva_res_1n}
\alias{netva_res_1n}
\title{Output of netva function as dataframe for all proteins in the breast cancer PPIN from single node approach.}
\description{
Output of netva function for all proteins in the breast cancer PPIN from single node approach
}
\details{
An dataframe with 10597 rows (proteins) and 14 columns (topological properties).
}
\references{
netva_res_1n = netva(vl, bc.net, ncore = 30).
}
\keyword{output-datasets}
