% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{netva_res_2n}
\alias{netva_res_2n}
\title{Output of netva function as dataframe for all VIPs in the breast cancer PPIN from two nodes approach.}
\description{
Output of netva function for all VIPs in the breast cancer PPIN from two nodes approach
}
\details{
An dataframe with 25838 rows (proteins pairs) and 14 columns (topological properties).
}
\references{
netva_res_2n = netva(vl, bc.net, ncore = 30).
}
\keyword{output-datasets}
