% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{netva_res_3n}
\alias{netva_res_3n}
\title{Output of netva function as dataframe for all VIPs in the breast cancer PPIN from three nodes approach.}
\description{
Output of netva function for all VIPs in the breast cancer PPIN from three nodes approach
}
\details{
An dataframe with 251270 rows (protein triplets) and 14 columns (topological properties).
}
\references{
netva_res_3n = netva(vl, bc.net, ncore = 30).
}
\keyword{output-datasets}
