% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectHub-Bot.R
\name{detectBottlenecks}
\alias{detectBottlenecks}
\title{Identify all possible bottlenecks based on pareto principle of 80:20 rule (by default) for a given network}
\usage{
detectBottlenecks(net, p = 20)
}
\arguments{
\item{net}{An igraph graph object.}

\item{p}{Value of percent of nodes/proteins to be considered for the determination of betweenness cutoff. By default value is 20\%. It is an optional parameter.}
}
\value{
A numeric vector containing all possible bottlenecks with bottleneck proteins' names and corresponding betweenness values as identified in the given network.
}
\description{
Identify all possible bottlenecks based on pareto principle of 80:20 rule (by default) for a given network
}
