% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectHub-Bot.R
\name{detectHubs}
\alias{detectHubs}
\title{Identify all possible hubs based on pareto principle of 80:20 rule (by default) for a given network}
\usage{
detectHubs(net, p = 20)
}
\arguments{
\item{net}{An igraph graph object.}

\item{p}{Value of percent of nodes/proteins to be considered for the determination of degree cutoff. By default value is 20\%. It is an optional parameter.}
}
\value{
A numeric vector containing all possible hubs with hub proteins' names and corresponding degree values as identified in the given network.
}
\description{
Identify all possible hubs based on pareto principle of 80:20 rule (by default) for a given network
}
