% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectVNs.R
\name{detectVNs}
\alias{detectVNs}
\title{Detects vulnerable nodes (VNs) based on five-point summary of boxplot}
\usage{
detectVNs(v, x, p)
}
\arguments{
\item{v}{Character vector containing names for all nodes of the given network.}

\item{x}{Numeric vector containing values of one topological property for all nodes of the given network. Where, the length of
v and x should be equal. Note: v and x should be in the same order i.e. position of one particular node in v and position of value of
topological property for that node in x should be the same.}

\item{p}{Case sensitive. Keyword for topological property based on which users want to identify VNs. Keyword should be one from the
following keywords: ANC (Average node connectivity), ND (Network density), NC (Network centralization), APL (Average path length), BET
(Betweenness), AP (Articulation point), NDR (Network diameter), CC (Clustering coefficient), and HET (Heterogeneity).}
}
\value{
A character vector containing all possible VNs.
}
\description{
This function detects vulnerable nodes (VNs) based on five-point summary of boxplot.
It takes first input a numeric vector of topological values for all nodes; whereas,
second a keyword for topological property based on which users want to identify VNs.
Different keywords for nine topological properties implemented in NetVA current version
are as follows: ANC (Average node connectivity), ND (Network density), NC (Network
centralization), APL (Average path length), BET (Betweenness), AP (Articulation
point), NDR (Network diameter), CC (Clustering coefficient), and HET (Heterogeneity).
}
