% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netva.R
\name{heterogeneity}
\alias{heterogeneity}
\title{Calculates value of heterogeneity for a given network}
\usage{
heterogeneity(net)
}
\arguments{
\item{net}{An igraph graph object.}
}
\value{
A single numeric value.
}
\description{
Calculates value of heterogeneity for a given network
}
