% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netva.R
\name{netva}
\alias{netva}
\title{Perform network vulnerability analysis for a given network}
\usage{
netva(v, net, type)
}
\arguments{
\item{v}{Name of one node or protein to be deleted}

\item{net}{Takes input an igraph graph object}

\item{type}{A keyword either "full" or "cluster" for consideration of complete network or only the largest cluster}
}
\value{
A dataframe of one row and ten columns
}
\description{
This function performs deletion of nodes one by one and calculation of average node connectivity, average path length, articulation
point, network centralization, network density, network diameter, transitivity (clustering coefficient), betweenness and heterogeneity
of resultant networks.
}
